Restore error count attributes in ErrorCountingFilter
